Refactored Event Manager
