+++ src/Event.c
@@ -7,18 +7,20 @@
 
 struct Event 
 {
-    void (*fun_ptr_arr[MAX_LISTNERS])(void); // 8*16 = 128
-    uint8_t currentRegisteredListeners; // 8
+    void (*fun_ptr_arr[MAX_LISTNERS])(void); 
+    uint8_t currentRegisteredListeners;
 };
 
 #define MAX_EVENTS 32
 
-static struct Event mEventStorage[MAX_EVENTS]; // 136 * 32 = 4,352 ?
-static EventPtr mEventPointers[MAX_EVENTS]; // 8*32
+static struct Event mEventStorage[MAX_EVENTS];
+static EventPtr mEventPointers[MAX_EVENTS];
 static uint8_t mNextUnusedEvent;
 
-// allocate some space to store deleted
-static size_t pointerStorage;
+static bool Event_internal_getElementIndex(EventPtr event, void (*ptr)(void), int* index);
+static void Event_internal_wipeAllListeners(EventPtr event);
+
+
 
 void Events_system_init()
 {
@@ -32,10 +34,10 @@ void Events_system_init()
 EventPtr Event_createNew()
 {
     if (mNextUnusedEvent >= MAX_EVENTS)
-    return 0; // no more free events
+        return 0; 
 
     EventPtr evPtr = mEventPointers[mNextUnusedEvent];
-    Event_wipeAllListener(evPtr);
+    Event_internal_wipeAllListeners(evPtr);
     mNextUnusedEvent++;
     return evPtr;
 }
@@ -43,31 +45,34 @@ EventPtr Event_createNew()
 bool Event_destroy(EventPtr event)
 {
     if (mNextUnusedEvent <= 0)
-    return false;
+        return false;
 
     mNextUnusedEvent--;
     mEventPointers[mNextUnusedEvent] = event;
     return true;
 }
 
-uint32_t Event_addListener(EventPtr event, void (*ptr)(void))
+uint8_t Event_addListener(EventPtr event, void (*ptr)(void))
 {
-    if(event->currentRegisteredListeners >= MAX_LISTNERS) return -1;
+    if(event->currentRegisteredListeners >= MAX_LISTNERS) 
+        return -1;
 
     event->fun_ptr_arr[event->currentRegisteredListeners++] = ptr;
     return 0;
 }
 
-uint32_t Event_removeListener(EventPtr event, void (*ptr)(void))
+uint8_t Event_removeListener(EventPtr event, void (*ptr)(void))
 {
-    int i = Event_internal_getElementIndex(event,ptr);
-    if(i==-1)return -1; // hop out if element doesnt exist
+    int i;
+    if(!Event_internal_getElementIndex(event,ptr,&i))
+        return -1;
+    
     event->fun_ptr_arr[i] = (NULL);
     
     for (size_t i = 0; i < event->currentRegisteredListeners; i++)
     {
         if(event->fun_ptr_arr[i] != NULL) continue;
-        if(i+1 < event->currentRegisteredListeners) // end of array
+        if(i+1 < event->currentRegisteredListeners) 
         {
             event->fun_ptr_arr[i] = event->fun_ptr_arr[i+1];
             event->fun_ptr_arr[i+1] = NULL;
@@ -75,29 +80,6 @@ uint32_t Event_removeListener(EventPtr event, void (*ptr)(void))
     }
     event->currentRegisteredListeners--; 
     return 0;
-}
-
-void Event_wipeAllListener(EventPtr event)
-{
-    if(event->currentRegisteredListeners == 0)return;
-    //
-    for (size_t i = 0; i < event->currentRegisteredListeners; i++)
-    {
-        event->fun_ptr_arr[i] = NULL;
-    }
-    event->currentRegisteredListeners = 0;
-}
-
-uint32_t Event_internal_getElementIndex(EventPtr event, void (*ptr)(void))
-{
-    for (size_t i = 0; i < event->currentRegisteredListeners; i++)
-    {
-        if(event->fun_ptr_arr[i] == ptr)
-        {
-            return i;
-        }
-    }
-    return -1;
 }
 
 void Event_invoke(EventPtr event)
@@ -108,9 +90,29 @@ void Event_invoke(EventPtr event)
     }
 }
 
-
 uint8_t Event_getTotalEvents()
 {
     return mNextUnusedEvent;
 }
 
+void Event_internal_wipeAllListeners(EventPtr event)
+{
+    for (size_t i = 0; i < event->currentRegisteredListeners; i++)
+    {
+        event->fun_ptr_arr[i] = NULL;
+    }
+    event->currentRegisteredListeners = 0;
+}
+
+static bool Event_internal_getElementIndex(EventPtr event, void (*ptr)(void), int* index)
+{
+    for (size_t i = 0; i < event->currentRegisteredListeners; i++)
+    {
+        if(event->fun_ptr_arr[i] == ptr)
+        {
+            *index = i;
+            return true;
+        }
+    }
+    return false;
+}
