+++ test/CMakeLists.txt
@@ -4,7 +4,10 @@ set(This ExampleTests)
 
 set(TestSources Tests.cpp)
 
-add_executable(${This} ${TestSources} ${Sources})
+set(TestMethods test_src/Methods.c
+                test_src/Methods.h)
+
+add_executable(${This} ${TestSources} ${Sources} ${TestMethods})
 
 target_link_libraries(${This} PUBLIC gtest_main)
